customer management is ok.

+++ src/main/java/org/kodluyoruz/mybank/customer/CustomerService.java
@@ -1,5 +1,6 @@
 package org.kodluyoruz.mybank.customer;
 
+import org.kodluyoruz.mybank.debit.DebitRepo;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.jpa.repository.Modifying;
@@ -11,9 +12,11 @@ import java.util.UUID;
 @Service
 public class CustomerService {
     private final CustomerRepo customerRepo;
+    private final DebitRepo debitRepo;
 
-    public CustomerService(CustomerRepo customerRepo) {
+    public CustomerService(CustomerRepo customerRepo, DebitRepo debitRepo) {
         this.customerRepo = customerRepo;
+        this.debitRepo = debitRepo;
     }
 
     public Customer create(Customer customer){
@@ -34,6 +37,9 @@ public class CustomerService {
     }
 
     public Customer delete(UUID id){
-        return customerRepo.delete(id);
+        if (customerRepo.findCustomerByAccounts_Empty() && debitRepo.findDebitByMinDebitIsNull()){
+            return customerRepo.delete(id);
+        }
+        return null;
     }
 }
